\name{JSONLines-package}
\alias{JSONLines-package}
\alias{JSONLines}
\docType{package}
\title{\packageTitle{JSONLines}}
\description{\packageDescription{JSONLines}}
\section{Package Content}{\packageIndices{JSONLines}}
\author{\packageAuthor{JSONLines}}
\section{Maintainer}{\packageMaintainer{JSONLines}}
\keyword{package}
