% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSONLines.R
\name{JSONLines_setup}
\alias{JSONLines_setup}
\title{Setup JSONLines}
\usage{
JSONLines_setup(...)
}
\arguments{
\item{...}{Parameters passed to JuliaCall::julia_setup}
}
\description{
First time setup to initialize Julia and the JSONLines package
}
\examples{

\donttest{

JSONLines::JSONLines_setup()

}
}
