% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSONLines.R
\name{readfile}
\alias{readfile}
\title{Read a JSONLines file}
\usage{
readfile(path, nrows = .Machine$integer.max, skip = 0)
}
\arguments{
\item{path}{Path to JSONLines file}

\item{nrows}{Number of rows to load}

\item{skip}{Number of rows to skip before loading}
}
\description{
Read a JSONLines file
}
