% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSONLines.R
\name{writefile}
\alias{writefile}
\title{Write a JSONLines file}
\usage{
writefile(x, file)
}
\arguments{
\item{x}{the data.frame to be written}

\item{file}{character string naming a file}
}
\description{
Write a JSONLines file
}
